+++ DatadogCore/Tests/Datadog/RUM/Integrations/CrashReportReceiverTests.swift
@@ -402,6 +402,7 @@ class CrashReportReceiverTests: XCTestCase {
         DDAssertReflectionEqual(sendRUMViewEvent.os, lastRUMViewEvent.os)
     }
 
+    //
     func testGivenCrashDuringRUMSessionWithActiveView_whenSendingRUMErrorEvent_itIsLinkedToPreviousRUMSessionAndIncludesCrashInformation() throws {
         let lastRUMViewEvent: RUMViewEvent = .mockRandomWith(crashCount: 0)
 
@@ -549,6 +550,175 @@ class CrashReportReceiverTests: XCTestCase {
         // Then
         let sendRUMErrorEvent = featureScope.eventsWritten(ofType: RUMCrashEvent.self)[0]
         XCTAssertEqual(sendRUMErrorEvent.model.error.sourceType, .iosIl2cpp, "Must send overridden sourceType")
+    }
+
+    func testGivenCrashDuringRUMSessionWithActiveViewAndEventMapper_whenSendingRUMViewEvent_itSendsMappedEvents() throws {
+        // Given
+        let lastRUMViewEvent: RUMViewEvent = .mockRandomWith(crashCount: 0)
+        let crashDate: Date = .mockDecember15th2019At10AMUTC()
+        let dateCorrectionOffset: TimeInterval = .mockRandom()
+        let crashReport: DDCrashReport = .mockWith(date: crashDate)
+        let crashContext: CrashContext = .mockWith(
+            serverTimeOffset: dateCorrectionOffset,
+            trackingConsent: .granted,
+            lastRUMViewEvent: AnyCodable(lastRUMViewEvent) // means there was a RUM session and it was sampled
+        )
+
+        let modifiedViewName = String.mockRandom()
+        let errorFingerprint = String.mockRandom()
+        let receiver: CrashReportReceiver = .mockWith(
+            featureScope: featureScope,
+            dateProvider: RelativeDateProvider(using: crashDate),
+            sessionSampler: Bool.random() ? .mockKeepAll() : .mockRejectAll(), // no matter sampling (as previous session was sampled)
+            trackBackgroundEvents: .mockRandom(), // no matter BET
+            eventsMapper: .mockWith(
+                viewEventMapper: { event in
+                    var event = event
+                    event.view.name = modifiedViewName
+                    return event
+                },
+                errorEventMapper: { event in
+                    var event = event
+                    event.error.fingerprint = errorFingerprint
+                    return event
+                }
+            )
+        )
+
+        // When
+        XCTAssertTrue(
+            receiver.receive(message: .baggage(
+                key: MessageBusSender.MessageKeys.crash,
+                value: MessageBusSender.Crash(report: crashReport, context: crashContext)
+            ), from: NOPDatadogCore())
+        )
+
+        // Then
+        let sendRUMViewEvent = featureScope.eventsWritten(ofType: RUMViewEvent.self).last
+        XCTAssertEqual(sendRUMViewEvent?.view.name, modifiedViewName, "Must send event mapper modified view event")
+
+        let sendRUMErrorEvent = featureScope.eventsWritten(ofType: RUMCrashEvent.self)[0]
+        XCTAssertEqual(sendRUMErrorEvent.model.error.fingerprint, errorFingerprint, "Must send event mapper modified error event")
+    }
+
+    func testGivenCrashDuringRUMSessionWithActiveView_whenErrorMapperReturnsNull_itSendOriginalError() throws {
+        let lastRUMViewEvent: RUMViewEvent = .mockRandomWith(crashCount: 0)
+
+        // Given
+        let crashDate: Date = .mockDecember15th2019At10AMUTC()
+        let crashReport: DDCrashReport = .mockWith(
+            date: crashDate,
+            type: "SIG_CODE (SIG_NAME)",
+            message: "Signal details",
+            stack: """
+            0: stack-trace line 0
+            1: stack-trace line 1
+            2: stack-trace line 2
+            """,
+            threads: [
+                .init(name: "Thread 0", stack: "thread 0 stack", crashed: true, state: nil),
+                .init(name: "Thread 1", stack: "thread 1 stack", crashed: false, state: nil),
+                .init(name: "Thread 2", stack: "thread 2 stack", crashed: false, state: nil),
+            ],
+            binaryImages: [
+                .init(libraryName: "library1", uuid: "uuid1", architecture: "arch", isSystemLibrary: true, loadAddress: "0xLoad1", maxAddress: "0xMax1"),
+                .init(libraryName: "library2", uuid: "uuid2", architecture: "arch", isSystemLibrary: true, loadAddress: "0xLoad2", maxAddress: "0xMax2"),
+                .init(libraryName: "library3", uuid: "uuid3", architecture: "arch", isSystemLibrary: false, loadAddress: "0xLoad3", maxAddress: "0xMax3"),
+            ],
+            meta: .init(
+                incidentIdentifier: "incident-identifier",
+                process: "process [1]",
+                parentProcess: "parent-process [0]",
+                path: "process/path",
+                codeType: "arch",
+                exceptionType: "EXCEPTION_TYPE",
+                exceptionCodes: "EXCEPTION_CODES"
+            )
+        )
+
+        let dateCorrectionOffset: TimeInterval = .mockRandom(min: 1, max: 5)
+        let crashContext: CrashContext = .mockWith(
+            serverTimeOffset: dateCorrectionOffset,
+            trackingConsent: .granted,
+            lastRUMViewEvent: AnyCodable(lastRUMViewEvent) // means there was a RUM session and it was sampled
+        )
+
+        let receiver: CrashReportReceiver = .mockWith(
+            featureScope: featureScope,
+            dateProvider: RelativeDateProvider(
+                using: crashDate.addingTimeInterval(
+                    .mockRandom(min: 10, max: 2 * CrashReportReceiver.Constants.viewEventAvailabilityThreshold) // simulate restarting app from 10s to 8h later
+                )
+            ),
+            sessionSampler: Bool.random() ? .mockKeepAll() : .mockRejectAll(), // no matter sampling (as previous session was sampled)
+            trackBackgroundEvents: .mockRandom(), // no matter BET
+            eventsMapper: .mockWith(
+                errorEventMapper: { _ in
+                    return nil
+                }
+            )
+        )
+
+        // When
+        XCTAssertTrue(
+            receiver.receive(message: .baggage(
+                key: MessageBusSender.MessageKeys.crash,
+                value: MessageBusSender.Crash(report: crashReport, context: crashContext)
+            ), from: NOPDatadogCore())
+        )
+
+        // Then
+        let sendRUMErrorEvent = featureScope.eventsWritten(ofType: RUMCrashEvent.self)[0]
+
+        XCTAssertTrue(
+            sendRUMErrorEvent.model.application.id == lastRUMViewEvent.application.id
+            && sendRUMErrorEvent.model.session.id == lastRUMViewEvent.session.id
+            && sendRUMErrorEvent.model.view.id == lastRUMViewEvent.view.id,
+            "The `RUMErrorEvent` sent must be linked to the same RUM Session as the last `RUMViewEvent`."
+        )
+
+        XCTAssertEqual(sendRUMErrorEvent.model.view.name, lastRUMViewEvent.view.name, "It must include view attributes")
+        XCTAssertEqual(sendRUMErrorEvent.model.view.referrer, lastRUMViewEvent.view.referrer, "It must include view attributes")
+        XCTAssertEqual(sendRUMErrorEvent.model.view.url, lastRUMViewEvent.view.url, "It must include view attributes")
+
+        XCTAssertNotNil(sendRUMErrorEvent.context, "It must contain context details")
+        XCTAssertNotNil(lastRUMViewEvent.context, "It must contain context details")
+
+        DDAssertJSONEqual(
+            AnyEncodable(sendRUMErrorEvent.context?.contextInfo),
+            AnyEncodable(lastRUMViewEvent.context?.contextInfo),
+            "The `RUMErrorEvent` sent must be include the context of the last `RUMViewEvent`."
+        )
+
+        XCTAssertTrue(
+            sendRUMErrorEvent.model.error.isCrash == true, "The `RUMErrorEvent` sent must be marked as crash."
+        )
+        XCTAssertEqual(
+            sendRUMErrorEvent.model.date,
+            crashDate.addingTimeInterval(dateCorrectionOffset).timeIntervalSince1970.toInt64Milliseconds,
+            "The `RUMErrorEvent` sent must include crash date corrected by current correction offset."
+        )
+        XCTAssertEqual(
+            sendRUMErrorEvent.model.error.type,
+            "SIG_CODE (SIG_NAME)"
+        )
+        XCTAssertEqual(
+            sendRUMErrorEvent.model.error.message,
+            "Signal details"
+        )
+        XCTAssertEqual(
+            sendRUMErrorEvent.model.error.stack,
+            """
+            0: stack-trace line 0
+            1: stack-trace line 1
+            2: stack-trace line 2
+            """
+        )
+        XCTAssertEqual(sendRUMErrorEvent.model.dd.session?.plan, .plan1, "All RUM events should use RUM Lite plan")
+        DDAssertJSONEqual(AnyEncodable(sendRUMErrorEvent.additionalAttributes?[DDError.threads]), crashReport.threads)
+        DDAssertJSONEqual(AnyEncodable(sendRUMErrorEvent.additionalAttributes?[DDError.binaryImages]), crashReport.binaryImages)
+        DDAssertJSONEqual(AnyEncodable(sendRUMErrorEvent.additionalAttributes?[DDError.meta]), crashReport.meta)
+        DDAssertJSONEqual(AnyEncodable(sendRUMErrorEvent.additionalAttributes?[DDError.wasTruncated]), crashReport.wasTruncated)
     }
 
     // MARK: - Testing Uploaded Data - Crashes During RUM Session With No Active View
@@ -767,6 +937,87 @@ class CrashReportReceiverTests: XCTestCase {
         )
     }
 
+    func testGivenCrashDuringRUMSessionWithNoActiveViewAndEventMapper_whenSendingRUMViewEvent_itSendsMappedEvents() throws {
+        func test(
+            lastRUMSessionState: RUMSessionState,
+            launchInForeground: Bool,
+            backgroundEventsTrackingEnabled: Bool
+        ) throws {
+            let featureScope = FeatureScopeMock()
+
+            // Given
+            let crashDate: Date = .mockDecember15th2019At10AMUTC()
+            let crashReport: DDCrashReport = .mockWith(
+                date: crashDate,
+                type: .mockRandom()
+            )
+            let dateCorrectionOffset: TimeInterval = .mockRandom()
+            let crashContext: CrashContext = .mockWith(
+                serverTimeOffset: dateCorrectionOffset,
+                service: .mockRandom(),
+                version: .mockRandom(),
+                buildNumber: .mockRandom(),
+                source: .mockRandom(),
+                trackingConsent: .granted,
+                userInfo: .mockRandom(),
+                networkConnectionInfo: .mockRandom(),
+                carrierInfo: .mockRandom(),
+                lastRUMViewEvent: nil, // means there was no active RUM view
+                lastRUMSessionState: AnyCodable(lastRUMSessionState), // means there was RUM session (sampled)
+                lastIsAppInForeground: launchInForeground
+            )
+
+            let mappedViewName = String.mockRandom()
+            let errorFingerprint = String.mockRandom()
+            let receiver: CrashReportReceiver = .mockWith(
+                featureScope: featureScope,
+                applicationID: .mockRandom(among: .alphanumerics),
+                dateProvider: RelativeDateProvider(using: crashDate),
+                sessionSampler: Bool.random() ? .mockKeepAll() : .mockRejectAll(), // no matter sampling (as previous session was sampled)
+                trackBackgroundEvents: backgroundEventsTrackingEnabled,
+                eventsMapper: .mockWith(
+                    viewEventMapper: { event in
+                        var event = event
+                        event.view.name = mappedViewName
+                        return event
+                    },
+                    errorEventMapper: { event in
+                        var event = event
+                        event.error.fingerprint = errorFingerprint
+                        return event
+                    }
+                )
+            )
+
+            // When
+            XCTAssertTrue(
+                receiver.receive(message: .baggage(
+                    key: MessageBusSender.MessageKeys.crash,
+                    value: MessageBusSender.Crash(report: crashReport, context: crashContext)
+                ), from: NOPDatadogCore())
+            )
+
+            // Then
+            let sentRUMView = featureScope.eventsWritten(ofType: RUMViewEvent.self).last
+            XCTAssertEqual(sentRUMView?.view.name, mappedViewName, "Must send mapped RUMView event")
+
+            let sentRUMError = featureScope.eventsWritten(ofType: RUMCrashEvent.self)[0]
+            XCTAssertEqual(sentRUMError.model.error.fingerprint, errorFingerprint, "Must send mapped RUMError event")
+        }
+
+        try test(
+            lastRUMSessionState: .mockWith(isInitialSession: true, hasTrackedAnyView: false), // when initial session with no views history
+            launchInForeground: true, // launch in foreground
+            backgroundEventsTrackingEnabled: .mockRandom() // no matter BET
+        )
+
+        try test(
+            lastRUMSessionState: .mockRandom(), // any sampled session
+            launchInForeground: false, // launch in background
+            backgroundEventsTrackingEnabled: true // BET enabled
+        )
+    }
+
     // MARK: - Testing Uploaded Data - Crashes During App Launch
 
     func testGivenCrashDuringAppLaunch_whenSending_itIsSendAsRUMErrorInNewRUMSession() throws {
@@ -947,6 +1198,109 @@ class CrashReportReceiverTests: XCTestCase {
             let sentRUMError = featureScope.eventsWritten(ofType: RUMCrashEvent.self)[0]
 
             XCTAssertEqual(sentRUMError.model.error.sourceType, .iosIl2cpp, "Must send overridden sourceType")
+        }
+
+        try test(
+            launchInForeground: true, // launch in foreground
+            backgroundEventsTrackingEnabled: .mockRandom(), // no matter BET
+            expectViewName: RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewName,
+            expectViewURL: RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewURL
+        )
+
+        try test(
+            launchInForeground: false, // launch in background
+            backgroundEventsTrackingEnabled: true, // BET enabled
+            expectViewName: RUMOffViewEventsHandlingRule.Constants.backgroundViewName,
+            expectViewURL: RUMOffViewEventsHandlingRule.Constants.backgroundViewURL
+        )
+    }
+
+    func testGivenCrashDuringAppLaunchWithEventMappers_whenMapperReturnsNil_itSendsOriginalEvents() throws {
+        func test(
+            launchInForeground: Bool,
+            backgroundEventsTrackingEnabled: Bool,
+            expectViewName expectedViewName: String,
+            expectViewURL expectedViewURL: String
+        ) throws {
+            let featureScope = FeatureScopeMock()
+
+            // Given
+            let crashDate: Date = .mockDecember15th2019At10AMUTC()
+            let crashReport: DDCrashReport = .mockWith(
+                date: crashDate,
+                type: .mockRandom()
+            )
+
+            let dateCorrectionOffset: TimeInterval = .mockRandom()
+            let randomNetworkConnectionInfo = NetworkConnectionInfo.mockRandom()
+            let randomCarrierInfo = CarrierInfo.mockRandom()
+            let randomUserInfo = UserInfo.mockRandom()
+            let crashContext: CrashContext = .mockWith(
+                serverTimeOffset: dateCorrectionOffset,
+                trackingConsent: .granted,
+                userInfo: randomUserInfo,
+                networkConnectionInfo: randomNetworkConnectionInfo,
+                carrierInfo: randomCarrierInfo,
+                lastRUMViewEvent: nil, // means there was no RUM session (it crashed during app launch)
+                lastRUMSessionState: nil, // means there was no RUM session (it crashed during app launch)
+                lastIsAppInForeground: launchInForeground
+            )
+
+            let mappedViewName = String.mockRandom()
+            let receiver: CrashReportReceiver = .mockWith(
+                featureScope: featureScope,
+                applicationID: .mockRandom(),
+                dateProvider: RelativeDateProvider(using: crashDate),
+                trackBackgroundEvents: backgroundEventsTrackingEnabled,
+                eventsMapper: .mockWith(
+                    viewEventMapper: { event in
+                        var event = event
+                        event.view.name = mappedViewName
+                        return event
+                    },
+                    errorEventMapper: { event in
+                        return nil
+                    }
+                )
+            )
+
+            // When
+            XCTAssertTrue(
+                receiver.receive(message: .baggage(
+                    key: MessageBusSender.MessageKeys.crash,
+                    value: MessageBusSender.Crash(report: crashReport, context: crashContext)
+                ), from: NOPDatadogCore())
+            )
+
+            // Then
+            let sentRUMView = featureScope.eventsWritten(ofType: RUMViewEvent.self).last
+            XCTAssertEqual(sentRUMView?.view.name, mappedViewName, "Must send mapped RUM view event")
+
+            // Sent RUM Error should be unmapped
+            let sentRUMError = featureScope.eventsWritten(ofType: RUMCrashEvent.self)[0]
+            XCTAssertNil(sentRUMError.model.error.fingerprint)
+            DDAssertReflectionEqual(
+                sentRUMError.model.connectivity,
+                RUMConnectivity(networkInfo: randomNetworkConnectionInfo, carrierInfo: randomCarrierInfo),
+                "It must contain connectity info from the moment of crash"
+            )
+            DDAssertJSONEqual(
+                sentRUMError.usr,
+                RUMUser(userInfo: randomUserInfo),
+                "It must contain user info from the moment of crash"
+            )
+            XCTAssertTrue(sentRUMError.model.error.isCrash == true, "RUM error must be marked as crash.")
+            XCTAssertEqual(
+                sentRUMError.model.date,
+                crashDate.addingTimeInterval(dateCorrectionOffset).timeIntervalSince1970.toInt64Milliseconds,
+                "RUM error must include crash date corrected by current correction offset."
+            )
+            XCTAssertEqual(sentRUMError.model.dd.session?.plan, .plan1, "All RUM events should use RUM Lite plan")
+            XCTAssertNotNil(sentRUMError.additionalAttributes?[DDError.threads], "It must contain crash details")
+            XCTAssertNotNil(sentRUMError.additionalAttributes?[DDError.binaryImages], "It must contain crash details")
+            XCTAssertNotNil(sentRUMError.additionalAttributes?[DDError.meta], "It must contain crash details")
+            XCTAssertNotNil(sentRUMError.additionalAttributes?[DDError.wasTruncated], "It must contain crash details")
+            XCTAssertEqual(sentRUMError.model.error.sourceType, .ios, "Must send .ios as the sourceType")
         }
 
         try test(